requirements.txt
Pillow>=9.0.0
requests>=2.0.0
beautifulsoup4>=4.0.0
